refactor: extract job polling, CLI update, and job execution logic into dedicated modules.
